feat: optimize key capacity (#364)

+++ camellia-redis-proxy/camellia-redis-proxy-core/src/main/java/com/netease/nim/camellia/redis/proxy/upstream/local/storage/key/block/KeyBlockReadWrite.java
@@ -68,10 +68,10 @@ public class KeyBlockReadWrite implements IKeyBlockReadWrite {
         }
         long fileId = slotInfo.fileId();
         long offset = slotInfo.offset();
-        int capacity = slotInfo.capacity();
-        int bucketSize = capacity / _4k;
+        long capacity = slotInfo.capacity();
+        int bucketSize = (int) (capacity / _4k);
         int bucket = KeyHashUtils.hash(key.key()) % bucketSize;
-        long bucketOffset = offset + bucket * _4k;
+        long bucketOffset = offset + (long) bucket * _4k;
 
         String cacheKey = fileId + "|" + bucketOffset;
         byte[] block = readCache.get(cacheKey);
