+++ camellia-redis-proxy/camellia-redis-proxy-core/src/main/java/com/netease/nim/camellia/redis/proxy/upstream/local/storage/key/slot/SlotInfo.java
@@ -3,6 +3,6 @@ package com.netease.nim.camellia.redis.proxy.upstream.local.storage.key.slot;
 /**
  * Created by caojiajun on 2024/12/31
  */
-public record SlotInfo(long fileId, long offset, int capacity) {
+public record SlotInfo(long fileId, long offset, long capacity) {
 
 }
